Update BundleUtilities and GetPrerequisites to the 2.8.3 versions.
We needed to patch BundleUtilities and GetPrerequisites to support
the deployment of Qt plugins and using package manager libraries
for our deployment. The two patches I've deleted were needed but
have now been merged into the CMake next branch so we're using the
versions of files from there. These files should be deleted when
2.8.3 is released and just the minimum version set of CMake set
to be 2.8.3.