Show week numbers in date pickers
That actually makes sense for every occurence of QDateEdit. Especially
when manually creating events inside the event editor window this is really helpful.

Note that this is disabled by default in QDateEdit (see usage of
QCalendarWidget::setVerticalHeaderFormat inside QCalendarPopup::verifyCalendarInstance()
inside qdatetimeedit.cpp). Not sure why, probably because the majority of users of
QDateEdit don't need that feature(?).