Fix crash in clearTasks() triggered on OS X
The clearTasks() apparently code relied on std::map::clear() to destroy
the contained TaskTreeItems in a certain order (probably in key order).
That caused a crash in CharmDataModelTests on OS X (10.9). Apparently
the clang/libstdc++ there doesn't clear in the assumed order, thus
the crash when accessing parents. Turns out there is already the right
fix in setAllTasks() (destroying the TaskTreeItem hierarchy before clearing),
so just unify clearTasks() and setAllTasks().