Adding the date time uploaded attribute for a timeSheet
LOCA wants to display the dateTimeUpload so the TimeSheetProcessor which
handle the db part need to store the info when inserting it. Plus I made
some test which uses the two convient new ctors from ComandLine. I added a test file within
Tests/Data folder to be used by the unit test.

Change-Id: I040dea8ea63c3b6c7413186ce4496b89e8058eca