Handle http basic auth
This makes use of QNM's authenticationRequired signal to hook
into http basic auth requests and pass on name+password.

Notes
1. the signal will not be emitted in case form-based logins
are used because no auth-required is send back to the client.
2. this is kind of a migration-strategy to keep the previous
way working as-is but in case we get a auth-required, like
happens with http basic auth enabled, we use that one the
and skip form login.
3. this way its possible to switch at the server-side the
auth method and keep the client transparent working.
4. price payed is an additional roundtrip, in case http
basic auth is enabled, which wouldn't be needed because
the credentials could also be embedded direct into the
request. Anyhow, this solution is the most less offensive
way to support both, form-login and http basic auth, for
the time being.