Fix Windows/Qt5 deployment
DeployQt4 does roughly the right thing also for Qt5 on
Windows, except the plugin deployment. Thus do that manually
(we only need platform and qsqlite plugin).

The issue of deploying Qt5 applications with CPack doesn't seem
to be properly solved yet (if I don't miss something). We might
want to fix that upstream or migrate the installer to something
else (windeployqt + CPack, Qt installer framework?) later.

For now, this fixes the immediate problem of deploying Charm on
Windows.

Fixes #200

Change-Id: I88b872f601fef72b3e4283ed366c2d49f9a8ddea
Reviewed-on: https://codereview.kdab.com/22499
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Daniel Kielack <daniel.kielack@kdab.com>
Reviewed-by: Guillermo Amaral <gamaral@kdab.com>