OS X: Prevent multiple "Check for updates" menu entries
We have three main windows all readding the same actions to the menus.
Check for updates is supposed to be in the app menu on OS X
(ApplicationSpecificRole). Unfortunately, setting that flag causes
up to three entries in the app menu. Thus move it to the File menu for
the time being.

Fixes #222

Change-Id: I4258565eb6d9f1b13800c2ce8b2bde948430b8d1
Reviewed-on: https://codereview.kdab.com/25793
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Boyer de la Giroday Michel <michel.giroday@kdab.com>
Reviewed-by: Frank Osterfeld <frank.osterfeld@kdab.com>