Fix httpclient error-message extraction
Entities like &nbsp; are invalid XML what leads
to QXmlStreamReader::hasError==true what aborts
parsing. Add a custom QXmlStreamEntityResolver
to just skip over such undeclared entities and
replace them with a whitespace. Thats enough
because we are only interested in keeping the
QXmlStreamReader parsing till we reached that
error-message we are supposed to extract.

Change-Id: Iacd722bb80e7b02df81e24e3951d9d45d6120575