Implement regular upload of staged timesheets
Upload a "staged" timesheet whenever a new task is started and no
staged timesheet was uploaded yet for the previous day. This gives
data for the last day, without uploading too often which would enable
the server to reconstruct the distribution of tracked time over the
day (which would be a privacy issue).

Change-Id: I8b46e9c2f79702ec264890f7a370ddaacc87e480
Reviewed-on: https://codereview.kdab.com/48931
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Steffen Hansen <steffen.hansen@kdab.com>
Reviewed-by: Kevin Krammer <kevin.krammer@kdab.com>