Un-minimize windows before raising them
When we call show() on a minimized widget, it would not un-minimize.
Do this manually by unsetting the minimized state. This allows us
to use the shortcuts to bring a dialog to the front even when it was
minimized in-between.

Note: showNormal would be an alternative, but that would unset the
maximized state, which may be undesireable to the user.