Un-minimize windows before raising them
When we call show() on a minimized widget, it would not un-minimize.
Do this manually by unsetting the minimized state. This allows us
to use the shortcuts to bring a dialog to the front even when it was
minimized in-between.

Note: showNormal would be an alternative, but that would unset the
maximized state, which may be undesireable to the user.

Change-Id: I86ee9fac7cb2413829b2c1e8de681569bfe21f77
Reviewed-on: https://codereview.kdab.com/53432
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Frank Osterfeld <frank.osterfeld@kdab.com>