Raise and show the window when staring charm and it's running already
This allows us to simply run charmtimetracker from the CLI or krunner
and it will always ensure the window is being shown. Currently, it
would only be shown on the first run. But when we minimize the window
afterwards and the app is running in the background, consecutive runs
of the command would not do anything besides printing an unhelpful
message. Now this behavior is much better.

Note: On KWin at least, this does not work when focus prevention
is set to anything but None. I guess this is due to lack of input-
taking widgets in the time tracker window. Anyhow, better than nothing
as it now at least works when the focus prevention is set to None.