Fix memory usage increasing over time, due to QActions piling up.
QMenu::clear deletes the actions if they have the menu as parent (that
is the case) and if they are not in any other widget (that isn't the
case, they are also in the systray menu). So when we finally remove the
actions from the systray, it's time to delete them -- and since the
removing there isn't using QMenu::clear, we need to do it by hand.

Testcase: qDebug() << m_menu->actions().count() at the end of
TimeTrackingTaskSelector::populate was showing 7, 24, 40, 56, 72, 88...
going up every time that tracking on a code is started.

Found with heaptrack (after noticing that a long running charm uses a
lot more memory than a freshly started one)

Change-Id: I8d6570ede5c569d400dd3fc63ef3789dbf13fb49
Reviewed-on: https://codereview.kdab.com/61730
Tested-by: Continuous Integration <build@kdab.com>
Reviewed-by: Filipe Azevedo <filipe.azevedo@kdab.com>